requirements.txt
langchain~=0.3.8
openai~=1.55.0
huggingface_hub
python-dotenv~=1.0.1
streamlit
torch~=2.5.0
transformers~=4.45.2
datasets~=3.0.1
gradio~=4.44.1
numpy~=1.26.4
plotly~=5.24.1
langchain-openai~=0.2.9
scikit-learn~=1.5.2
beautifulsoup4~=4.12.3
langchain-community~=0.3.8
langchain-core~=0.3.21
langchain-text-splitters~=0.3.2
rdflib~=7.1.1